Implements authentication Context
